Move SIV out of Miscreant::Internals and make internals private
This makes the SIV class public as Miscreant::SIV, and marks the
Miscreant::Internals module as private, hiding it from the outside world.

Miscreant::AEAD and Miscreant::SIV are now the only exposed APIs.
