+++ ruby/lib/miscreant.rb
@@ -6,6 +6,7 @@ require "securerandom"
 
 require "miscreant/version"
 require "miscreant/aead"
+require "miscreant/siv"
 require "miscreant/internals"
 
 # Miscreant: A misuse-resistant symmetric encryption library
@@ -15,4 +16,7 @@ module Miscreant
 
   # Ciphertext failed to verify as authentic
   IntegrityError = Class.new(CryptoError)
+
+  # Hide internals from the outside world
+  private_constant :Internals
 end
