+++ ruby/spec/spec_helper.rb
@@ -3,6 +3,14 @@
 
 require "bundler/setup"
 require "miscreant"
-require "support/test_vectors"
 
 RSpec.configure(&:disable_monkey_patching!)
+
+# Un-hide the hidden Internals constant so we can test the internals
+module Miscreant
+  internals = Internals
+  remove_const :Internals
+  const_set :Internals, internals
+end
+
+require "support/test_vectors"
